requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
build==1.2.1
CacheControl==0.14. 
celery==5.3.5
certifi==2024.2.2
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.8
Django==5.0.4
django-cors-headers==3.9.0
django-split-settings==1.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.14.0
idna==3.7
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
kombu==5.3.7
more-itertools==10.2.0
msgpack==1.0.8
packaging==24.0
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.1
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.8.1
requests==2.31.0
requests-toolbelt==1.0.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sqlparse==0.5.0
tomlkit==0.12.4
trove-classifiers==2024.4.10
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.1
wcwidth==0.2.13
